requirements.txt
langchain
openai
tiktoken
python-dotenv
websockets
requests
streamlit